Integrate extended document symbols into DocumentSymbolHandler.
- Automatically include import package, and type declaration elements
  into the outline
- Modify the symbol details to exclude return type, but continue to
  include fully qualified name for fields/methods
