+++ org.eclipse.jdt.ls.core/src/org/eclipse/jdt/ls/core/internal/handlers/JDTLanguageServer.java
@@ -56,7 +56,7 @@ import org.eclipse.jdt.ls.core.internal.LanguageServerWorkingCopyOwner;
 import org.eclipse.jdt.ls.core.internal.ServiceStatus;
 import org.eclipse.jdt.ls.core.internal.codemanipulation.GenerateGetterSetterOperation.AccessorField;
 import org.eclipse.jdt.ls.core.internal.handlers.CodeActionHandler.CodeActionData;
-import org.eclipse.jdt.ls.core.internal.handlers.ExtendedDocumentSymbolHandler.ExtendedDocumentSymbol;
+import org.eclipse.jdt.ls.core.internal.handlers.DocumentSymbolHandler.ExtendedDocumentSymbol;
 import org.eclipse.jdt.ls.core.internal.handlers.ExtractInterfaceHandler.CheckExtractInterfaceResponse;
 import org.eclipse.jdt.ls.core.internal.handlers.FindLinksHandler.FindLinksParams;
 import org.eclipse.jdt.ls.core.internal.handlers.GenerateAccessorsHandler.AccessorCodeActionParams;
@@ -752,11 +752,11 @@ public class JDTLanguageServer extends BaseJDTLanguageServer implements Language
 
 	@Override
 	public CompletableFuture<List<ExtendedDocumentSymbol>> extendedDocumentSymbols(DocumentSymbolParams params) {
-		logInfo(">> document/extendedDocumentSymbols");
-		ExtendedDocumentSymbolHandler handler = new ExtendedDocumentSymbolHandler(preferenceManager);
+		logInfo(">> java/extendedDocumentSymbols");
+		DocumentSymbolHandler handler = new DocumentSymbolHandler(preferenceManager);
 		return computeAsync((monitor) -> {
 			waitForLifecycleJobs(monitor);
-			return handler.handle(params, monitor);
+			return handler.extendedDocumentSymbol(params, monitor);
 		});
 	}
